pooling operation for mitie featurizer.
