+++ rasa/nlu/featurizers/dense_featurizer/mitie_featurizer.py
@@ -1,6 +1,6 @@
 import numpy as np
 import typing
-from typing import Any, List, Text, Optional
+from typing import Any, List, Text, Optional, Dict
 
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.featurizers.featurizer import Featurizer
@@ -26,6 +26,17 @@ class MitieFeaturizer(Featurizer):
     requires = [TOKENS_NAMES[attribute] for attribute in MESSAGE_ATTRIBUTES] + [
         "mitie_feature_extractor"
     ]
+
+    defaults = {
+        # Specify what pooling operation should be used to calculate the vector of
+        # the CLS token. Available options: 'mean' and 'max'
+        "pooling": "mean"
+    }
+
+    def __init__(self, component_config: Optional[Dict[Text, Any]] = None):
+        super().__init__(component_config)
+
+        self.pooling_operation = self.component_config["pooling"]
 
     @classmethod
     def required_packages(cls) -> List[Text]:
@@ -107,7 +118,7 @@ class MitieFeaturizer(Featurizer):
             features.append(feature_extractor.get_feature_vector(token.text))
         features = np.array(features)
 
-        cls_token_vec = np.mean(features, axis=0, keepdims=True)
+        cls_token_vec = self._calculate_cls_vector(features, self.pooling_operation)
         features = np.concatenate([features, cls_token_vec])
 
         return features
