+++ tests/nlu/featurizers/test_featurizer.py
@@ -101,3 +101,26 @@ def test_sequence_to_sentence_features(features, expected):
         assert np.all(expected.toarray() == actual.toarray())
     else:
         assert np.all(expected == actual)
+
+
+@pytest.mark.parametrize(
+    "pooling, features, expected",
+    [
+        (
+            "mean",
+            np.array([[0.5, 3, 0.4, 0.1], [0, 0, 0, 0], [0.5, 3, 0.4, 0.1]]),
+            np.array([[0.5, 3, 0.4, 0.1]]),
+        ),
+        (
+            "max",
+            np.array([[1.0, 3.0, 0.0, 2.0], [4.0, 3.0, 1.0, 0.0]]),
+            np.array([[4.0, 3.0, 1.0, 2.0]]),
+        ),
+    ],
+)
+def test_calculate_cls_vector(pooling, features, expected):
+    featurizer = Featurizer()
+
+    actual = featurizer._calculate_cls_vector(features, pooling)
+
+    assert np.all(actual == expected)
